Chipmunk testbed adapt non-win platform compile issue

+++ tests/cpp-tests/Classes/ChipmunkTestBed/ChipmunkTestBed.cpp
@@ -97,7 +97,7 @@ cpConstraint* mouse_joint = NULL;
 char const* ChipmunkDemoMessageString = NULL;
 
 #define GRABBABLE_MASK_BIT (1 << 31)
-cpShapeFilter GRAB_FILTER = { CP_NO_GROUP, GRABBABLE_MASK_BIT, GRABBABLE_MASK_BIT };
+cpShapeFilter GRAB_FILTER = { CP_NO_GROUP, (unsigned int)GRABBABLE_MASK_BIT, (unsigned int)GRABBABLE_MASK_BIT };
 cpShapeFilter NOT_GRABBABLE_FILTER = { CP_NO_GROUP, ~GRABBABLE_MASK_BIT, ~GRABBABLE_MASK_BIT };
